+++ CHANGELOG.md
@@ -57,6 +57,7 @@
 ## [Unreleased][]
 ### Added
 ### Changed
+- UI: tag slots on search button can now be cloned. Note cloning also carries over any modification introduced by the active recipe.
 - UI: tag entries on info button can now be cloned.
 - Configuration: changed the remove duplicates bias to prefer tracks containing 'BEST' within a 'TRACKDSP' tag.
 - Helpers: updated helpers.